requirements/base.txt
Django==4.2.16
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
django-filter==24.3
djangorestframework-simplejwt==5.2.2
drf-case-middleware==0.1.2
Pillow==9.4.0
psycopg2-binary==2.9.6
requests==2.31.0
python-decouple==3.8
dj-rest-auth[with_social]==6.0.0